% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycadas-package.R
\docType{package}
\name{cycadas-package}
\alias{cycadas}
\alias{cycadas-package}
\title{cycadas: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Oliver Hunewald \email{oliver.hunewald@lih.lu} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
