% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{cycadas}
\alias{cycadas}
\title{Main function to run the Cycadas Shiny app}
\usage{
cycadas()
}
\description{
This function is the entry point for running the Cycadas Shiny app.
}
\details{
It calls the ShinyApp function, which initializes the Shiny app with the
specified UI and server components.
}
\seealso{
\code{\link{shinyApp}}
}
