% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{kmeansTH}
\alias{kmeansTH}
\title{Perform k-means clustering and identify bimodal expression thresholds}
\usage{
kmeansTH(df)
}
\arguments{
\item{df}{A dataframe of expression data.}
}
\value{
th: A dataframe containing the calculated thresholds for each column.
}
\description{
This function takes a dataframe of expression data where each column represents
an expression profile. It checks whether the data in each column is bimodally
distributed and calculates the threshold between high and low expression using
1D k-means clustering.
}
\details{
The function checks the bimodal distribution of each column in the input
dataframe. If a column is determined to be bimodal, the threshold separating
high and low expression is calculated using 1D k-means clustering.
}
